linkpull/requirements.txt
python-fasthtml
requests
beautifulsoup4
